Extract PyPI class for definition lookup
