+++ lib/license_finder/package_managers/pip.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'json'
-require 'net/http'
 
 module LicenseFinder
   class Pip < PackageManager
@@ -17,7 +16,7 @@ module LicenseFinder
         PipPackage.new(
           name,
           version,
-          pypi_def(name, version),
+          PyPI.pypi_def(name, version),
           logger: logger,
           children: children,
           install_path: Pathname(location).join(name)
@@ -61,20 +60,6 @@ module LicenseFinder
       JSON(output).map do |package|
         package.values_at('name', 'version', 'dependencies', 'location')
       end
-    end
-
-    def pypi_def(name, version)
-      response = pypi_request("https://pypi.org/pypi/#{name}/#{version}/json")
-      response.is_a?(Net::HTTPSuccess) ? JSON.parse(response.body).fetch('info', {}) : {}
-    end
-
-    def pypi_request(location, limit = 10)
-      uri = URI(location)
-      http = Net::HTTP.new(uri.host, uri.port)
-      http.use_ssl = true
-      response = http.get(uri.request_uri).response
-
-      response.is_a?(Net::HTTPRedirection) && limit.positive? ? pypi_request(response['location'], limit - 1) : response
     end
   end
 end
